Unused import and other tidying.

+++ src/gwt/client/mail/MailPage.java
@@ -36,7 +36,7 @@ public class MailPage extends Page
                 } else if (extra.equals("i")) {
                     compose.setGiftItem((byte)args.get(++ii, 0), args.get(++ii, 0));
                 } else if (extra.equals("r")) {
-                    compose.setGiftRoom((int)args.get(++ii, 0));
+                    compose.setGiftRoom(args.get(++ii, 0));
                 }
             }
             setContent(_msgs.mailTitle(), compose);
